SimpleChatTC: Move console.log trapping into its own module
So that it can be used from different modules, if required.
